added code to update profile image after update

+++ app/Http/Controllers/API/UserController.php
@@ -4,9 +4,8 @@ namespace App\Http\Controllers\API;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Hash;
-
 use App\User;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -19,7 +18,6 @@ class UserController extends Controller
     {
         $this->middleware('auth:api');
     }
-
     /**
      * Display a listing of the resource.
      *
@@ -27,9 +25,11 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // $this->authorize('isAdmin');
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            return User::latest()->paginate(5);
+        }
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -39,20 +39,47 @@ class UserController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'name' => ['required', 'min:3'],
-            'email' => ['required', 'min:3', 'unique:users'],
-            'password' => ['required', 'min:6',],
-            'name' => ['required', 'min:3'],
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users',
+            'password' => 'required|string|min:6'
         ]);
-        User::create([
-            'name' => request('name'),
-            'email' => request('email'),
-            'password' => Hash::make(request('password')),
-            'type' => request('type'),
-            'bio' => request('bio'),
+        return User::create([
+            'name' => $request['name'],
+            'email' => $request['email'],
+            'type' => $request['type'],
+            'bio' => $request['bio'],
+            'photo' => $request['photo'],
+            'password' => Hash::make($request['password']),
         ]);
     }
-
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,' . $user->id,
+            'password' => 'sometimes|required|min:6'
+        ]);
+        $currentPhoto = $user->photo;
+        if ($request->photo != $currentPhoto) {
+            $name = time() . '.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
+            \Image::make($request->photo)->save(public_path('img/profile/') . $name);
+            $request->merge(['photo' => $name]);
+            $userPhoto = public_path('img/profile/') . $currentPhoto;
+            if (file_exists($userPhoto)) {
+                @unlink($userPhoto);
+            }
+        }
+        if (!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+        $user->update($request->all());
+        return ['message' => "Success"];
+    }
+    public function profile()
+    {
+        return auth('api')->user();
+    }
     /**
      * Display the specified resource.
      *
@@ -63,20 +90,6 @@ class UserController extends Controller
     {
         //
     }
-
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function profile()
-    {
-        //
-        return auth('api')->user();
-    }
-
     /**
      * Update the specified resource in storage.
      *
@@ -86,20 +99,15 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
-        $user  =  User::findorFail($id);
-        //validate input
+        $user = User::findOrFail($id);
         $this->validate($request, [
-            'name' => ['required', 'min:3'],
-            'email' => 'required|email|min:3|unique:users,email,' . $user->id,
-            'password' => ['sometimes', 'min:6',],
-            'name' => ['required', 'min:3'],
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,' . $user->id,
+            'password' => 'sometimes|min:6'
         ]);
-
         $user->update($request->all());
-        return ["message" => "User update"];;
+        return ['message' => 'Updated the user info'];
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -108,10 +116,22 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        //find user
-        $user  =  User::findorFail($id);
-        //delete user
+        $this->authorize('isAdmin');
+        $user = User::findOrFail($id);
+        // delete the user
         $user->delete();
-        return ["message" => "User deleted"];
+        return ['message' => 'User Deleted'];
+    }
+    public function search()
+    {
+        if ($search = \Request::get('q')) {
+            $users = User::where(function ($query) use ($search) {
+                $query->where('name', 'LIKE', "%$search%")
+                    ->orWhere('email', 'LIKE', "%$search%");
+            })->paginate(20);
+        } else {
+            $users = User::latest()->paginate(5);
+        }
+        return $users;
     }
 }
